refactor: split release/rollback command modules (#223)
Extract the bump-type resolution helpers from release.py into
release_versioning.py (-> 606 lines) and the git tag/commit plumbing from
rollback.py into rollback_git.py (-> 535 lines). Both new modules are
re-exported from their parents, so the public import surface is unchanged.
Test patch targets are updated to where the moved code now resolves its
dependencies (release_versioning.* / rollback_git.run_git_command), and the
module-size gate is tightened from 800 to 750 to lock in the smaller band.
